--- mapillary_tools/geotag/__init__.py
@@ -1,1 +1,0 @@
-from .. import geo  # noqa: F401
