+++ mapillary_tools/geotag/geotag_images_from_gpx_file.py
@@ -4,9 +4,7 @@ import logging
 import typing as T
 from pathlib import Path
 
-import gpxpy
-
-from .. import geo
+from . import utils
 from .geotag_images_from_gpx import GeotagImagesFromGPX
 
 
@@ -23,7 +21,7 @@ class GeotagImagesFromGPXFile(GeotagImagesFromGPX):
         num_processes: int | None = None,
     ):
         try:
-            tracks = parse_gpx(source_path)
+            tracks = utils.parse_gpx(source_path)
         except Exception as ex:
             raise RuntimeError(
                 f"Error parsing GPX {source_path}: {ex.__class__.__name__}: {ex}"
@@ -43,30 +41,3 @@ class GeotagImagesFromGPXFile(GeotagImagesFromGPX):
             offset_time=offset_time,
             num_processes=num_processes,
         )
-
-
-Track = T.List[geo.Point]
-
-
-def parse_gpx(gpx_file: Path) -> list[Track]:
-    with gpx_file.open("r") as f:
-        gpx = gpxpy.parse(f)
-
-    tracks: list[Track] = []
-
-    for track in gpx.tracks:
-        for segment in track.segments:
-            tracks.append([])
-            for point in segment.points:
-                if point.time is not None:
-                    tracks[-1].append(
-                        geo.Point(
-                            time=geo.as_unix_time(point.time),
-                            lat=point.latitude,
-                            lon=point.longitude,
-                            alt=point.elevation,
-                            angle=None,
-                        )
-                    )
-
-    return tracks
