+++ mapillary_tools/geotag/geotag_images_from_video.py
@@ -5,8 +5,7 @@ import typing as T
 from pathlib import Path
 
 from .. import types, utils
-
-from .geotag_from_generic import GeotagImagesFromGeneric
+from .base import GeotagImagesFromGeneric
 from .geotag_images_from_gpx import GeotagImagesFromGPX
